refactor: change class name and category
