+++ src/main/java/racingcar/controller/Controller.java
@@ -1,26 +1,26 @@
 package racingcar.controller;
 
 import java.util.List;
-import racingcar.domain.game.RacingGame;
+import racingcar.domain.Game;
 import racingcar.domain.random.RandomNumberGenerator;
 import racingcar.view.View;
 
 public class Controller {
 
     private final View view;
-    private final RacingGame racingGame;
+    private final Game game;
 
     public Controller(List<String> names, int tryCount, RandomNumberGenerator randomNumberGenerator) {
         this.view = new View();
-        this.racingGame = new RacingGame(names, tryCount, randomNumberGenerator);
+        this.game = new Game(names, tryCount, randomNumberGenerator);
     }
 
     public void runGame() {
         view.printResultViewTitle();
-        while (!racingGame.isFinished()) {
-            racingGame.proceedTurn();
-            view.printMidtermResults(racingGame.getGameLog(), racingGame.getTryCount().getCurrentTryCount());
+        while (!game.isFinished()) {
+            game.proceedTurn();
+            view.printMidtermResults(game.getGameLog(), game.getTryCount().getCurrentTryCount());
         }
-        view.printWinnersResult(racingGame.getGameLog(), racingGame.getTryCount().getTotalTryCount());
+        view.printWinnersResult(game.getGameLog(), game.getTryCount().getTotalTryCount());
     }
 }
