+++ src/main/java/racingcar/domain/Game.java
@@ -1,4 +1,4 @@
-package racingcar.domain.game;
+package racingcar.domain;
 
 import java.util.List;
 import racingcar.domain.Cars;
@@ -7,14 +7,14 @@ import racingcar.domain.log.GameLog;
 import racingcar.domain.random.RandomNumberGenerator;
 import racingcar.exception.RacingGameIsFinishedException;
 
-public class RacingGame {
+public class Game {
 
     private final Cars cars;
     private final TryCount tryCount;
     private final RandomNumberGenerator randomNumberGenerator;
     private final GameLog gameLog;
 
-    public RacingGame(List<String> names, int tryCount, RandomNumberGenerator randomNumberGenerator) {
+    public Game(List<String> names, int tryCount, RandomNumberGenerator randomNumberGenerator) {
         this.cars = new Cars(names);
         this.tryCount = new TryCount(tryCount);
         this.randomNumberGenerator = randomNumberGenerator;
